Add the `SsspFamily` subclass of `Group` (#8)
Subclassing of the `Group` entity is a new feature that ships with
`aiida-core==1.2.0`. We use that here to create a subclass that can be
used to define a group of `UpfData` nodes. By making this an official
class, the creation, validation and manipulation of it is made a lot
easier. The logic to construct a family from a directory of pseudos can
now be implemented on this class, instead of having to provide a stand
alone function. Additionally, the validation can be implemented on the
class, for example to guarantee that there is only a single pseudo for
each element. Finally, the class implements an internal cache to make
the looking up of pseudos lazy and as efficient as possible.