Add the `SsspParameters` data plugin (#12)
This class is intended to host the parameters that are tied to a
particular instance of a `SsspFamily`. The dictionary contains
for a set of elements a dictionary with the filename and md5 of
the corresponding pseudo in the family and the recommended wave
function and charge density cutoff. Note that the validation that of
whether the correct elements are present and the validity of the
filename and md5 checksum, is left up to the caller.