Adapt `aiida-sssp install` to the new SSSP Archive entry
The previous version of the Materials Cloud Archive entry for the SSSP
provided files with inconsistent format and it did not provide the files
for all available SSSP configurations. For example, PBEsol was not
accessible from the Archive, which is however supposed to be the
canonical source.

A new version of the entry was created:

    https://archive.materialscloud.org/2018.0001/v4

which now provides download links for the pseudo archive and JSON
metadata file for all available SSSP configurations. The archives and
metadata files also now follow a strict and well-defined structure and
naming convention, making it a lot easier for programmatic solutions
like `aiida-sssp install`.