`SsspFamily`: fix bug that rendered it broken when loaded from database
The initialization of the class depended on the constructor being
called, however, when loading the `Group` subclass instance from the
database through the ORM, the constructor is not called. To fix this,
the initialization of required attributes is done by defining them as
class attributes and the `pseudos` property is added to lazily load
them when requested.