`SsspParameters`: fully decouple from the `SsspFamily` class
The `SsspParameters` depended indirectly on the `SsspFamily` class by
requiring an instance of the latter to be passed in to the constructor.
However, this would lead to a circular dependency, because some
factories for the `SsspFamily` class would also ideally be able to
construct an instance of `SsspParameters`.

Since it makes more sense to have the flow of control stem from the
family, we remove the dependency by simply allowind to pass in a string
label in the constructor of `SsspParameters`. It is then up to the
family to make sure that the metadata of a parameters node is
commensurate with its actual pseudo potentials.