`SsspParameters`: use UUID of `SsspFamily` instead of label to join
The use of the label of the `SsspFamily` to connect an `SsspParameters`
to it was a historical solution, but, given that a `Group` label is not
guaranteed to be unique and, worse, it is mutable, makes it a horrible
attribute for a joining attribute. Therefore, we change it to use the
UUID of the `SsspFamily` which is both unique and immutable, addressing
both limitations of the label.