Generalize the `SsspFamily` to the `UpfFamily` group plugin
The functionality provided by the `SsspFamily` is very useful but can
currently, by design choice, only be used for actual SSSP configurations.
The original idea of the plugin and `SsspFamily` is to not make it
easier for peopl to work with the SSSP but also make it less likely that
incorrect SSSP configurations are used.

However, there is still a common use case for UPF families that are not
strict SSSP configurations. Therefore, we generalize it to the new
plugin `UpfFamily` which represents a generic family of UPF files. The
`SsspFamily` then simply is a subclass as it just imposes additional
constraints.

The idea is that the `SsspFamily` can only be used for SSSP
configurations and all other use cases will employ the `UpfFamily`
plugin instead. Note that it will be difficult to prevent people from
using the `SsspFamily` group for non-SSSP families, but there will never
be a way to prevent this. Instead we will simply not provide a command
line interface to create `SsspFamily` instances from anything other than
the official resources from the Materials Cloud Archive entry.